+++ src/services/common/common.js
@@ -34,3 +34,11 @@ export const saveTokens = (accessToken, refreshToken) => {
   setCookie('accessToken', accessToken);
   localStorage.setItem('refreshToken', refreshToken);
 };
+
+export const removeCookie = (sKey, sPath, sDomain) => {
+  document.cookie =
+    encodeURIComponent(sKey) +
+    '=; expires=Thu, 01 Jan 1970 00:00:00 GMT' +
+    (sDomain ? '; domain=' + sDomain : '') +
+    (sPath ? '; path=' + sPath : '');
+};
