+++ src/services/reducers/rootReducer.js
@@ -7,7 +7,6 @@ import { registrationReducer } from './registration';
 import { userReducer } from './user';
 import { forgotPasswordReducer } from './forgot-password';
 import { passwordResetReducer } from './password-reset';
-import { authReducer } from './auth';
 
 export const rootReducer = combineReducers({
   ingredients: ingredientsReducer,
@@ -18,5 +17,4 @@ export const rootReducer = combineReducers({
   user: userReducer,
   forgotPassword: forgotPasswordReducer,
   passwordReset: passwordResetReducer,
-  auth: authReducer,
 });
